backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
fpdf2==2.7.6
matplotlib==3.8.2
numpy==1.24.3
llama-cpp-python==0.2.11